refactor: Break styled components into smaller pieces and change imports accordingly
