+++ src/components/Layout.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { FullWidthContainer, ContentContainer } from './atoms/LayoutContainers';
+import FullWidthContainer from './atoms/FullWidthContainer';
+import ContentContainer from './atoms/ContentContainer';
 
 const Layout: React.SFC<{ children: JSX.Element[] }> = props => {
   return (
