+++ src/components/atoms/ContentContainer.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const ContentContainer = styled.div`
+  width: 85%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`;
+
+export default ContentContainer;
