+++ src/components/atoms/FullWidthContainer.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const FullWidthContainer = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export default FullWidthContainer;
