--- src/components/atoms/LayoutContainers.tsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const FullWidthContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const ContentContainer = styled.div`
-  width: 85%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-`;
