+++ src/components/atoms/Table.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const Table = styled.table`
+  border-collapse: collapse;
+  align-self: center;
+  max-width: 400px;
+  min-width: 250px;
+  margin-bottom: 1rem;
+`;
+
+Table.displayName = 'styledTable';
+
+export default Table;
