+++ src/components/atoms/Title.tsx
@@ -4,10 +4,12 @@ type TitleProps = {
   theme: any;
 };
 
-export const Title = styled.h1<TitleProps>`
+const Title = styled.h1<TitleProps>`
   text-align: center;
   color: ${props => props.theme.primary};
   font-weight: bold;
 `;
 
 Title.displayName = 'Title';
+
+export default Title;
